+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next release
+
+- support default value for `Enums` in Dart code generation.
+
 ## 0.2.0 (04-Dec-2023)
 
 - improved `Floats` to accept any `num`.